cleaned code, removed helpers, added plot for video pinball
